+++ packages/apalis-core/src/lib.rs
@@ -144,7 +144,6 @@ pub mod test_utils {
     use std::sync::atomic::{AtomicBool, Ordering};
     use std::sync::Arc;
     use std::task::{Context, Poll};
-    use std::time::Duration;
     use tower::{Layer, Service, ServiceBuilder};
 
     /// Define a dummy service
@@ -296,11 +295,20 @@ pub mod test_utils {
         /// Gets the current state of results
         pub async fn execute_next(&mut self) -> Option<(TaskId, Result<String, String>)> {
             self.should_next.store(true, Ordering::Release);
-            let res = futures::future::select(
-                self.res_rx.next(),
-                crate::sleep(Duration::from_secs(2)).boxed(),
-            )
-            .await;
+            #[cfg(feature = "sleep")]
+            let fut = async {
+                crate::sleep(std::time::Duration::from_secs(2))
+                    .boxed()
+                    .await;
+            }
+            .boxed();
+            #[cfg(not(feature = "sleep"))]
+            let fut = async {
+                std::future::pending::<()>().await;
+            }
+            .boxed();
+
+            let res = futures::future::select(self.res_rx.next(), fut).await;
             match res {
                 Either::Left(next) => next.0,
                 Either::Right(_) => None,