+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project are documented in this file.
 
-## [Unreleased]
+## [0.7.0](https://github.com/geofmureithi/apalis/releases/tag/v0.7.0)
 
 ### Added
 - **api** add associated types to the Backend trait ([#516](https://github.com/geofmureithi/apalis/pull/516))
@@ -11,6 +11,7 @@ All notable changes to this project are documented in this file.
 - **native TLS**: Add `async-std-comp-native-tls` and `tokio-comp-native-tls` features ([#525](https://github.com/geofmureithi/apalis/pull/525))
 - **cron** : Introduce CronContext ([#488](https://github.com/geofmureithi/apalis/pull/488))
 - **stepped tasks** : adds ability to execute stepped tasks ([#478](https://github.com/geofmureithi/apalis/pull/478))
+- **SQL** : add support for job priority to SQL storages ([#533](https://github.com/geofmureithi/apalis/pull/533/))
 
 ### Fixed
 - **PostgresStorage**: PostgresStorage get_jobs status conditional ([#524](https://github.com/geofmureithi/apalis/pull/524))
@@ -19,6 +20,7 @@ All notable changes to this project are documented in this file.
 - **orphaned tasks**: Re-enqueue orphaned jobs before starting streaming ([#507](https://github.com/geofmureithi/apalis/pull/507))
 - **deps**: Update Rust crate `redis` to `0.28` ([#495](https://github.com/geofmureithi/apalis/pull/495))
 - **deps**: Update Rust crate `redis` to `0.29` and `deadpool-redis` to `0.20` ([#527](https://github.com/geofmureithi/apalis/pull/527))
+- **features**: fix: ease apalis-core default features ([538](https://github.com/geofmureithi/apalis/pull/538))
 
 ### Tests
 - **Integration tests**: Aborting jobs and panicking workers ([#508](https://github.com/geofmureithi/apalis/pull/508))